chores(mcp): move mcp index into root index
